Fix Viewanima.php

+++ Viewanima.php
@@ -51,7 +51,7 @@ $result1 = $db->query($sql1);
   </div>
   <div class="col-sm-4 ">
     <p>If you wish to adopt this animal or want to learn more contact it's current foster. If you run into issues be sure to <a href="Contactus.php">contact us.</p> 
-    <?php if ($_SESSION['login_role'] == 'admin' || $_SESSION['login_id'] == $row['creator_id'] ){
+    <?php if (isset($_SESSION['login_role']) && $_SESSION['login_role'] == 'admin' || isset($_SESSION['login_role']) && $_SESSION['login_id'] == $row['creator_id'] ){
      echo "<a href='deleteanima.php?id=".$row['id']."'>Delete</a>";
   }?>
 </div>
